+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Update to latest Citizens. This is NOT backwards compatible! :(
  - Fix very short range of custom sound effects
+ - Fix NPC "caster" parameter
 
 ## 5.1.3